prototipo-vta/backend/requirements.txt
Flask
psycopg2-binary
python-dotenv
Werkzeug
reportlab==4.0.9
